Move error bound classes to separate file
